+++ src/CmsCoreServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Idoneo\CmsCore;
 
 use Idoneo\CmsCore\Commands\CreateAdminCommand;
+use Idoneo\CmsCore\Commands\DiagnoseCommand;
 use Idoneo\CmsCore\Commands\GenerateApiToken;
 use Idoneo\CmsCore\Commands\InstallCommand;
 use Idoneo\CmsCore\Commands\UpdateCommand;
@@ -48,6 +49,7 @@ class CmsCoreServiceProvider extends ServiceProvider
 				UpdateCommand::class,
 				CreateAdminCommand::class,
 				GenerateApiToken::class,
+				DiagnoseCommand::class,
 			]);
 		}
 	}